Added ability to add in any HTML into a DIV
This commit isn't completely related to inserting code. I had this rather complicated table (full of hyper-links, formatting, etc...) that I needed to display on multiple slides. I didn't want to have to cut-and-paste copies of the table everywhere, and seeing all that html in the slide made it a bit un-clean to navigate through the slides for editing. So, I added an extra feature to this library to just insert in content from an html file into a div.